+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `odood server wait-pg` command
 - Added `odood server run --wait-pg` option
 - Added `odood db is-initialized` command to check if database is already initialized.
+- Added `odood db ensure-initialized` command to  initialize database if it is not initialized yet.
 
 ### Changed